Crate serious

Serious is a simple language for concise mathematical expressions.

The numerical type is f64 (infinities and NaNs yield errors).

Variables are identified by characters within [A-Za-z].

Multiplication is implicit where an operator is omitted.

All operations are infix binary, except for the unary minus.

Operations are left-associative unless overridden by parentheses or precedence rules:

Operator	Meaning	Precedence

^	Exponentiate	2

*	Multiply	1

/	Divide	1

+	Add	0

-	Subtract	0

Example Usage:

use serious::{create_context, interpreter::interpret};

// create a context of bound variables
let (x, y) = (12.34, 9999.);
let context = create_context!{'x' => x, 'y' => y};

// evaluate an expression, given the context
let result = interpret("34.2x + y^2(-2x^3 + 1)/5.2", &context).unwrap();
assert_eq!(result, 34.2*x + y.powf(2.)*(-2.*x.powf(3.) + 1.)/5.2);

Modules

error
Defines the type for Result::Errs of parse, interpret, and interpret_tree.

interpreter
Evaluates an Expression, given a Context of bound identifiers.

parser
Converts text into an Expression (an abstract syntax tree).

Macros

create_context
Creates a Context which can be applied to an expression using interpret/interpret_tree.