Crate serious

Serious is a simple language to evaluate concise mathematical expressions.

The numerical type is f64 (infinities and NaNs raise errors).

Variables are identified by characters within [A-Za-z].

Multiplication is implicit if an operator is omitted, unless the RHS is a constant.

3x means 3*x.

x3 is a parse error.

All operations are infix binary, except for the unary minus.

Operations are left-associative unless overridden by parentheses or precedence rules:

Operator	Meaning	Precedence

^	Exponentiate	2

*	Multipy	1

/	Divide	1

+	Add	0

-	Subtract	0

Example Usage:

use serious::{create_context, interpret};

// create a context of bound variables
let (x, y) = (12.34, 9999.);
let context = create_context!{'x' => x, 'y' => y};

// evaluate an expression, given the context
let result = interpret("y^2(-2x^3 + 1)/5.2", &context).unwrap();
assert_eq!(result, y.powf(2.)*(-2.*x.powf(3.) + 1.)/5.2);

Re-exports

pub use parser::parse;

pub use interpreter::Context;

pub use interpreter::interpret;

pub use interpreter::interpret_tree;

pub use error::Error;

pub use error::ErrorType;

Modules

error
Defines the type for expressions that fail to evaluate.

interpreter
Evaluates an abstract syntax, given a context of bound identifiers.

parser
Converts tokens into an abstract syntax tree.

Macros

create_context
Creates a serious::Context which can be applied to an expression using serious::interpret.